Posting Description Preview - Now Functional
I have made the preview section functional by saving a line item
description into an object and saving the key for each description as
the ID for the listItem. This key value object is saved in local
storage so the descriptions save as well. When 'Preview Post' is
clicked it iterates through the list of items and uses the keys to find
the values for the line item. Those are added line by line to the
preview section.

Also, I added the delivery options and Notes For Entire Post values to
the preview section.